+++ src/main/java/org/folio/rest/impl/LookupsUtils.java
@@ -1,68 +1,189 @@
 package org.folio.rest.impl;
 
-import io.vertx.core.json.JsonObject;
-import org.folio.patron.rest.exceptions.HttpException;
-import org.folio.rest.RestVerticle;
-import org.folio.rest.tools.client.HttpClientFactory;
-import org.folio.rest.tools.client.Response;
-import org.folio.rest.tools.client.interfaces.HttpClientInterface;
-import org.folio.rest.tools.utils.TenantTool;
-
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 
+import org.apache.commons.lang3.StringUtils;
+import org.folio.patron.rest.exceptions.HttpException;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.MultiMap;
+import io.vertx.core.Vertx;
+import io.vertx.core.buffer.Buffer;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.client.HttpResponse;
+import io.vertx.ext.web.client.WebClient;
+
 class LookupsUtils {
+  private LookupsUtils() {}
 
-  private LookupsUtils(){}
-
-  static CompletableFuture<JsonObject> getItem(String itemId, Map<String, String> okapiHeaders,
-                                               HttpClientInterface httpClient) {
-    return get("/inventory/items/" + itemId, httpClient, okapiHeaders)
+  static CompletableFuture<JsonObject> getItem(String itemId, Map<String, String> okapiHeaders) {
+    return get("/inventory/items/" + itemId, okapiHeaders)
       .thenApply(LookupsUtils::verifyAndExtractBody);
   }
 
-  static CompletableFuture<JsonObject> getUser(String userId, Map<String, String> okapiHeaders,
-                                               HttpClientInterface httpClient) {
-    return get("/users/" + userId, httpClient, okapiHeaders)
+  static CompletableFuture<JsonObject> getUser(String userId, Map<String, String> okapiHeaders) {
+    return get("/users/" + userId, okapiHeaders)
       .thenApply(LookupsUtils::verifyAndExtractBody);
   }
 
-  static CompletableFuture<JsonObject> getRequestPolicyId(String queryString, Map<String, String> okapiHeaders,
-                                                          HttpClientInterface httpClient) {
-    return get("/circulation/rules/request-policy?" + queryString, httpClient, okapiHeaders)
+  static CompletableFuture<JsonObject> getRequestPolicyId(RequestTypeParameters criteria, Map<String, String> okapiHeaders) {
+    final var queryParameters = Map.of(
+      "item_type_id", criteria.getItemMaterialTypeId(),
+      "loan_type_id", criteria.getItemLoanTypeId(),
+      "patron_type_id", criteria.getPatronGroupId(),
+      "location_id", criteria.getItemLocationId());
+
+    return get("/circulation/rules/request-policy",
+        queryParameters, okapiHeaders)
       .thenApply(LookupsUtils::verifyAndExtractBody);
   }
 
-  static CompletableFuture<JsonObject> getRequestPolicy(String requestPolicyId, Map<String, String> okapiHeaders,
-                                                        HttpClientInterface httpClient) {
-    return get("/request-policy-storage/request-policies/" + requestPolicyId, httpClient, okapiHeaders)
+  static CompletableFuture<JsonObject> getRequestPolicy(String requestPolicyId, Map<String, String> okapiHeaders) {
+    return get("/request-policy-storage/request-policies/" + requestPolicyId, okapiHeaders)
       .thenApply(LookupsUtils::verifyAndExtractBody);
   }
 
   static JsonObject verifyAndExtractBody(Response response) {
-    if (!Response.isSuccess(response.getCode())) {
-      throw new CompletionException(new HttpException(response.getCode(),
-        response.getError().getString("errorMessage")));
+    if (!response.isSuccess()) {
+      throw new CompletionException(new HttpException(response.statusCode,
+        response.body));
     }
 
-    return response.getBody();
+    // Parsing an emppty body to JSON causes an exception
+    if (StringUtils.isBlank(response.body)) {
+      return null;
+    }
+
+    return new JsonObject(response.body);
   }
 
-  static HttpClientInterface getHttpClient(Map<String, String> okapiHeaders) {
-    final String okapiURL = okapiHeaders.getOrDefault("X-Okapi-Url", "");
-    final String tenantId = TenantTool.calculateTenantId(okapiHeaders.get(RestVerticle.OKAPI_HEADER_TENANT));
+  public static CompletableFuture<LookupsUtils.Response> post(String path,
+    JsonObject body, Map<String, String> okapiHeaders) {
 
-    return HttpClientFactory.getHttpClient(okapiURL, tenantId);
-  }
+    Vertx vertx = Vertx.currentContext().owner();
+    URL url;
 
-  private static CompletableFuture<Response> get (String path,
-                                                        HttpClientInterface httpClient,
-                                                        Map<String, String> okapiHeaders) {
     try {
-      return httpClient.request(path, okapiHeaders);
-    } catch (Exception e) {
-      throw new CompletionException(e);
+      url = new URL(buildUri(path, okapiHeaders));
+    } catch (MalformedURLException e) {
+      throw new CompletionException(e.getCause());
+    }
+
+    final var futureResponse
+      = new CompletableFuture<AsyncResult<HttpResponse<Buffer>>>();
+
+    final var request = WebClient.create(vertx)
+      .post(url.getPort(), url.getHost(), url.getPath())
+      .putHeaders(buildHeaders(okapiHeaders));
+
+
+    request.sendJson(body, futureResponse::complete);
+
+    return futureResponse
+      .thenCompose(LookupsUtils::toResponse);
+  }
+
+  public static CompletableFuture<LookupsUtils.Response> put(String path,
+    JsonObject body, Map<String, String> okapiHeaders) {
+
+    Vertx vertx = Vertx.currentContext().owner();
+    URL url;
+
+    try {
+      url = new URL(buildUri(path, okapiHeaders));
+    } catch (MalformedURLException e) {
+      throw new CompletionException(e.getCause());
+    }
+
+    final var futureResponse
+      = new CompletableFuture<AsyncResult<HttpResponse<Buffer>>>();
+
+    final var request = WebClient.create(vertx)
+      .put(url.getPort(), url.getHost(), url.getPath())
+      .putHeaders(buildHeaders(okapiHeaders));
+
+    request
+      .timeout(1000)
+      .sendJson(body, futureResponse::complete);
+
+    return futureResponse
+      .thenCompose(LookupsUtils::toResponse);
+  }
+
+  public static CompletableFuture<LookupsUtils.Response> get(String path,
+   Map<String, String> queryParameters, Map<String, String> okapiHeaders) {
+
+    Vertx vertx = Vertx.currentContext().owner();
+    URL url;
+
+    try {
+      url = new URL(buildUri(path, okapiHeaders));
+    } catch (MalformedURLException e) {
+      throw new CompletionException(e.getCause());
+    }
+
+    final var futureResponse
+      = new CompletableFuture<AsyncResult<HttpResponse<Buffer>>>();
+
+    final var request = WebClient.create(vertx)
+      .get(url.getPort(), url.getHost(), url.getPath())
+      .putHeaders(buildHeaders(okapiHeaders));
+
+    queryParameters.forEach(request::addQueryParam);
+
+    request.send(futureResponse::complete);
+
+    return futureResponse
+      .thenCompose(LookupsUtils::toResponse);
+  }
+
+  public static CompletableFuture<LookupsUtils.Response> get(String path,
+    Map<String, String> okapiHeaders) {
+
+    return get(path, Map.of(), okapiHeaders);
+  }
+
+  private static CompletableFuture<LookupsUtils.Response> toResponse(
+    AsyncResult<HttpResponse<Buffer>> result) {
+
+    if (result.failed()) {
+      return CompletableFuture.failedFuture(result.cause());
+    }
+
+    final var response = result.result();
+
+    final var mappedResponse = new Response(response.statusCode(),
+      response.bodyAsString());
+
+    return CompletableFuture.completedFuture(mappedResponse);
+  }
+
+  private static String buildUri(String path, Map<String, String> okapiHeaders) {
+    final String okapiURL = okapiHeaders.getOrDefault("X-Okapi-Url", "");
+
+    return okapiURL + path;
+  }
+
+  private static MultiMap buildHeaders(Map<String, String> okapiHeaders) {
+    return MultiMap.caseInsensitiveMultiMap()
+      .addAll(okapiHeaders);
+  }
+
+  public static class Response {
+    public final int statusCode;
+    public final String body;
+
+    public Response(int statusCode, String body) {
+      this.statusCode = statusCode;
+      this.body = body;
+    }
+
+    public boolean isSuccess() {
+      return statusCode >= 200 && statusCode < 300;
     }
   }
 }