+++ src/main/java/org/folio/rest/impl/RequestObjectFactory.java
@@ -2,7 +2,6 @@ package org.folio.rest.impl;
 
 import io.vertx.core.json.JsonObject;
 import org.folio.rest.jaxrs.model.Hold;
-import org.folio.rest.tools.client.interfaces.HttpClientInterface;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 
@@ -19,8 +18,8 @@ class RequestObjectFactory {
     okapiHeaders = headers;
   }
 
-  CompletableFuture<JsonObject> createRequestByItem(String patronId, String itemId, Hold entity, HttpClientInterface httpClient) {
-    return getRequestType(patronId, itemId, httpClient)
+  CompletableFuture<JsonObject> createRequestByItem(String patronId, String itemId, Hold entity) {
+    return getRequestType(patronId, itemId)
       .thenApply(requestType -> {
         if (requestType != RequestType.NONE) {
           final JsonObject holdJSON = new JsonObject()
@@ -29,7 +28,8 @@ class RequestObjectFactory {
             .put("requestType", requestType.getValue())
             .put(Constants.JSON_FIELD_REQUEST_DATE, new DateTime(entity.getRequestDate(), DateTimeZone.UTC).toString())
             .put("fulfilmentPreference", Constants.JSON_VALUE_HOLD_SHELF)
-            .put(Constants.JSON_FIELD_PICKUP_SERVICE_POINT_ID, entity.getPickupLocationId());
+            .put(Constants.JSON_FIELD_PICKUP_SERVICE_POINT_ID, entity.getPickupLocationId())
+            .put(Constants.JSON_FIELD_PATRON_COMMENTS, entity.getPatronComments());
 
           if (entity.getExpirationDate() != null) {
             holdJSON.put(Constants.JSON_FIELD_REQUEST_EXPIRATION_DATE,
@@ -42,18 +42,18 @@ class RequestObjectFactory {
       });
   }
 
-  private CompletableFuture<RequestType> getRequestType(String patronId, String itemId, HttpClientInterface httpClient ) {
+  private CompletableFuture<RequestType> getRequestType(String patronId, String itemId) {
 
-    CompletableFuture<JsonObject> userFuture = LookupsUtils.getUser(patronId, okapiHeaders, httpClient);
-    CompletableFuture<JsonObject> itemFuture = LookupsUtils.getItem(itemId, okapiHeaders, httpClient);
+    CompletableFuture<JsonObject> userFuture = LookupsUtils.getUser(patronId, okapiHeaders);
+    CompletableFuture<JsonObject> itemFuture = LookupsUtils.getItem(itemId, okapiHeaders);
 
     RequestTypeParameters requestTypeParams = new RequestTypeParameters();
 
     return CompletableFuture.allOf(userFuture, itemFuture)
       .thenApply(x -> createRequestPolicyIdCriteria(itemFuture, userFuture, requestTypeParams))
-      .thenCompose((RequestTypeParameters criteria) -> lookupRequestPolicyId(criteria, httpClient))
+      .thenCompose(this::lookupRequestPolicyId)
       .thenCompose(policyIdResponse ->
-          LookupsUtils.getRequestPolicy(policyIdResponse.getString("requestPolicyId"), okapiHeaders, httpClient))
+          LookupsUtils.getRequestPolicy(policyIdResponse.getString("requestPolicyId"), okapiHeaders))
       .thenApply(RequestPolicy::from)
       .thenApply(requestPolicy -> getRequestType(requestPolicy, requestTypeParams.getItemStatus()));
   }
@@ -68,14 +68,8 @@ class RequestObjectFactory {
     return RequestType.NONE;
   }
 
-  private CompletableFuture<JsonObject> lookupRequestPolicyId(RequestTypeParameters criteria, HttpClientInterface httpClient) {
-    String queryString = String.format(
-      "item_type_id=%s&loan_type_id=%s&patron_type_id=%s&location_id=%s",
-      criteria.getItemMaterialTypeId(), criteria.getItemLoanTypeId(),
-      criteria.getPatronGroupId(), criteria.getItemLocationId()
-    );
-
-    return LookupsUtils.getRequestPolicyId(queryString, okapiHeaders, httpClient);
+  private CompletableFuture<JsonObject> lookupRequestPolicyId(RequestTypeParameters criteria) {
+    return LookupsUtils.getRequestPolicyId(criteria, okapiHeaders);
   }
 
   private RequestTypeParameters createRequestPolicyIdCriteria(CompletableFuture<JsonObject> itemFuture,